move packages around SourceInspector to repository
