+++ linen-modern/src/main/scala/x7c1/linen/modern/init/updater/SourceUpdaterQueue.scala
@@ -6,6 +6,7 @@ import android.app.Service
 import android.database.sqlite.SQLiteConstraintException
 import x7c1.linen.database.control.DatabaseHelper
 import x7c1.linen.glue.service.ServiceControl
+import x7c1.linen.repository.crawler.{LoadedSource, InspectedSource, SourceInspector}
 import x7c1.wheat.modern.formatter.ThrowableFormatter
 import ThrowableFormatter.format
 import x7c1.wheat.macros.logger.Log
@@ -18,6 +19,7 @@ class SourceUpdaterQueue(
   helper: DatabaseHelper){
 
   private lazy val inspector = SourceInspector(helper)
+
   private lazy val queueMap = new SourceQueueMap
 
   def enqueue(source: InspectedSource): Unit = synchronized {
