+++ linen-repository/src/main/scala/x7c1/linen/repository/crawler/InvalidEntry.scala
@@ -1,4 +1,4 @@
-package x7c1.linen.modern.init.updater
+package x7c1.linen.repository.crawler
 
 sealed trait InvalidEntry
